+++ ErsatzTV.FFmpeg/Capabilities/IHardwareCapabilitiesFactory.cs
@@ -6,7 +6,9 @@ public interface IHardwareCapabilitiesFactory
 {
     void ClearCache();
 
-    Task<IFFmpegCapabilities> GetFFmpegCapabilities(string ffmpegPath);
+    Task<IFFmpegCapabilities> GetFFmpegCapabilities(string ffmpegPath, CancellationToken cancellationToken);
+
+    Task<string> GetFFmpegVersion(string ffmpegPath, CancellationToken cancellationToken);
 
     Task<IHardwareCapabilities> GetHardwareCapabilities(
         IFFmpegCapabilities ffmpegCapabilities,